+++ je_auto_control/utils/governance/__init__.py
@@ -2,9 +2,9 @@
 from je_auto_control.utils.governance.credential_broker import (
     CredentialBroker, CredentialBrokerError, default_broker, set_secret_resolver,
 )
-from je_auto_control.utils.governance.governance import ApprovalGate
+from je_auto_control.utils.governance.governance import ApprovalGate, approval_gate
 
 __all__ = [
-    "ApprovalGate", "CredentialBroker", "CredentialBrokerError",
+    "ApprovalGate", "CredentialBroker", "CredentialBrokerError", "approval_gate",
     "default_broker", "set_secret_resolver",
 ]